Changed to Long type

+++ whiteboard/src/main/java/login/PersonRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 public interface PersonRepository extends Repository<Person, Long> {
 	
-	Person findById(@Param("id")int id);
+	Person findById(@Param("id")Long id);
 	void save(Person p);
